#47 refactor-before-option   - 옵션 구현 전 정리
