+++ src/main/java/com/project/eventlink/common/BasicData.java
@@ -1,31 +1,27 @@
 package com.project.eventlink.common;
 
-import com.project.eventlink.entity.Role;
-import com.project.eventlink.member.dto.JoinForm;
-import com.project.eventlink.member.service.MemberService;
-import jakarta.annotation.PostConstruct;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
+/**
+ * FIXME 초기 더미 데이터는 test의 data.sql
+ *  파일에 저장하는 것으로 대체하는 것이 좋을것 같습니다.
+ */
+//@Component
+//@RequiredArgsConstructor
 public class BasicData {
+//    private final MemberService memberService;
 
-    private final MemberService memberService;
-
-    @PostConstruct
-    public void init() {
-        JoinForm test1 = new JoinForm();
-        test1.setUserId("a");
-        test1.setPassword("1");
-        test1.setRole(Role.USER);
-
-        JoinForm test2 = new JoinForm();
-        test2.setUserId("b");
-        test2.setPassword("1");
-        test2.setRole(Role.PROVIDER);
-
-        memberService.save(test1);
-        memberService.save(test2);
-    }
+//    @PostConstruct
+//    public void init() {
+//        JoinForm test1 = new JoinForm();
+//        test1.setUserId("a");
+//        test1.setPassword("1");
+//        test1.setRole(Role.USER);
+//
+//        JoinForm test2 = new JoinForm();
+//        test2.setUserId("b");
+//        test2.setPassword("1");
+//        test2.setRole(Role.PROVIDER);
+//
+//        memberService.save(test1);
+//        memberService.save(test2);
+//    }
 }
