+++ src/main/resources/application-local.yml
@@ -5,7 +5,7 @@ spring:
 
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: validate
     generate-ddl: true
     open-in-view: false
     properties:
