+++ src/test/java/com/project/eventlink/EventLinkApplicationTests.java
@@ -3,10 +3,6 @@ package com.project.eventlink;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
-class EventLinkApplicationTests {
 
-    @Test
-    void contextLoads() {
-    }
+class EventLinkApplicationTests {
 }
